backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.2.0",
    "aiosqlite>=0.21.0",
    "chromadb>=0.6.3",
    "cryptography>=45.0.2",
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.3",
    "langchain-community>=0.3.24",
    "langchain-deepseek>=0.1.3",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "mysql-connector-python>=9.3.0",
    "pydantic-settings>=2.9.1",
    "pymysql>=1.1.1",
    "pypdf>=5.6.0",
    "pyppeteer>=2.0.0",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]>=0.34.2",
]
